+++ src/abilities/index.ts
@@ -1,2 +1,3 @@
 export * from './abilityDefinition.js';
+export * from './passiveAbilities.js';
 export * from './useAbility.js';